Replace schemes when injecting into app + refactor connections
