+++ src/shared/modules/app/appDuck.js
@@ -30,6 +30,9 @@ export const DESKTOP = 'DESKTOP'
 export const WEB = 'WEB'
 export const CLOUD = 'CLOUD'
 
+const SECURE_SCHEMES = ['neo4j+s', 'bolt+s']
+const INSECURE_SCHEMES = ['neo4j', 'bolt']
+
 // Selectors
 export const getHostedUrl = state => (state[NAME] || {}).hostedUrl || null
 export const getEnv = state => (state[NAME] || {}).env || WEB
@@ -37,14 +40,16 @@ export const hasDiscoveryEndpoint = state =>
   [WEB, CLOUD].includes(getEnv(state))
 export const inWebEnv = state => getEnv(state) === WEB
 export const inWebBrowser = state => [WEB, CLOUD].includes(getEnv(state))
-export const getAllowedBoltSchemes = state => {
+export const getAllowedBoltSchemes = (state, encryptionFlag) => {
   const isHosted = inWebBrowser(state)
   const hostedUrl = getHostedUrl(state)
   return !isHosted
-    ? null
-    : hostedUrl.startsWith('https')
-    ? ['neo4j+s', 'bolt+s']
-    : ['neo4j', 'bolt']
+    ? encryptionFlag
+      ? SECURE_SCHEMES
+      : INSECURE_SCHEMES
+    : (hostedUrl || '').startsWith('https')
+    ? SECURE_SCHEMES
+    : INSECURE_SCHEMES
 }
 
 // Reducer
