+++ src/shared/services/bolt/bolt.js
@@ -24,6 +24,12 @@ import WorkPool from '../WorkPool'
 import * as mappings from './boltMappings'
 import * as boltConnection from './boltConnection'
 import {
+  cancelTransaction as globalCancelTransaction,
+  routedReadTransaction as globalRoutedReadTransaction,
+  routedWriteTransaction as globalRoutedWriteTransaction,
+  directTransaction as globalDirectTransaction
+} from './transactions'
+import {
   runCypherMessage,
   cancelTransactionMessage,
   closeConnectionMessage
@@ -64,7 +70,7 @@ function cancelTransaction(id, cb) {
     work.onFinish(cb)
     work.execute(cancelTransactionMessage(id))
   } else {
-    boltConnection.cancelTransaction(id, cb)
+    globalCancelTransaction(id, cb)
   }
 }
 
@@ -101,7 +107,7 @@ function routedWriteTransaction(input, parameters, requestMetaData = {}) {
     )
     return [id, workerPromise]
   } else {
-    return boltConnection.routedWriteTransaction(input, parameters, {
+    return globalRoutedWriteTransaction(input, parameters, {
       requestId,
       cancelable,
       txMetadata,
@@ -142,7 +148,7 @@ function routedReadTransaction(input, parameters, requestMetaData = {}) {
     )
     return workerPromise
   } else {
-    return boltConnection.routedReadTransaction(input, parameters, {
+    return globalRoutedReadTransaction(input, parameters, {
       requestId,
       cancelable,
       txMetadata,
@@ -182,7 +188,7 @@ function directTransaction(input, parameters, requestMetaData = {}) {
     )
     return workerPromise
   } else {
-    return boltConnection.directTransaction(input, parameters, {
+    return globalDirectTransaction(input, parameters, {
       requestId,
       cancelable,
       txMetadata,
@@ -260,6 +266,5 @@ export default {
       intConverter: val => val.toNumber(),
       objectConverter: mappings.extractFromNeoObjects
     }),
-  neo4j,
   addTypesAsField
 }
